consolidated Character Card Style Imports to CharacterCardStyles.js
