+++ src/components/CharacterCard.js
@@ -1,10 +1,6 @@
 import React from "react";
-import Container from "@material-ui/core/Container";
-import CssBaseline from "@material-ui/core/CssBaseline";
-import Typography from "@material-ui/core/Typography";
-import Grid from "@material-ui/core/Grid";
-import makeStyles from "@material-ui/core/styles/makeStyles";
-import {Paper} from "@material-ui/core";
+
+import { Paper, Typography, Grid, makeStyles, CssBaseline, Container } from './Styles/CharacterCardStyles';
 
 const useStyles = makeStyles(theme => ({
     paper: {
