+++ src/components/Styles/CharacterCardStyles.js
@@ -1,0 +1,8 @@
+import CssBaseline from "@material-ui/core/CssBaseline";
+import Typography from "@material-ui/core/Typography";
+import Grid from "@material-ui/core/Grid";
+import makeStyles from "@material-ui/core/styles/makeStyles";
+import {Paper} from "@material-ui/core";
+import Container from "@material-ui/core/Container";
+
+export { Paper, Typography, Grid, makeStyles, CssBaseline, Container }
